feat: 1)upload image with type 'heic'  2)change the empty message to ';' for models  3)
